Clean up interface for PR
